PP-2396 Split transactions list controller into own file
* Split transactions list ctrl into separate file as it serves a distinct and separate purpose
